requirements.txt
google-genai>=0.3.0
pydantic>=2.0.0
requests>=2.28.0
networkx>=3.0
matplotlib>=3.6.0
